fix: avoid MFA Challenge if stepUp is not requiring it when the user is authenticated
fixes AM-3647

gravitee-io/issues#9908
