+++ gravitee-am-gateway/gravitee-am-gateway-handler/gravitee-am-gateway-handler-oidc/src/main/java/io/gravitee/am/gateway/handler/oauth2/resources/endpoint/authorization/AuthorizationEndpoint.java
@@ -37,6 +37,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import static io.gravitee.am.common.utils.ConstantKeys.ACTION_KEY;
+import static io.gravitee.am.common.utils.ConstantKeys.AUTH_FLOW_FINALIZED_KEY;
 
 /**
  * The authorization endpoint is used to interact with the resource owner and obtain an authorization grant.
@@ -91,6 +92,7 @@ public class AuthorizationEndpoint implements Handler<RoutingContext> {
                             try {
                                 // final step of the authorization flow, we can clean the session and redirect the user
                                 cleanSession(context);
+                                context.session().put(AUTH_FLOW_FINALIZED_KEY, true);
                                 doRedirect(context, request, authorizationResponse);
                             } catch (Exception e) {
                                 logger.error("Unable to redirect to client redirect_uri", e);
